Updates imports on tests
