+++ library/src/test/java/tmg/utilities/utils/LocalDateUtilsTest.kt
@@ -2,102 +2,81 @@ package tmg.utilities.utils
 
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 import org.threeten.bp.LocalDate
-import org.threeten.bp.LocalTime
 import org.threeten.bp.format.DateTimeParseException
-import tmg.utilities.extensions.toLocalDate
-import tmg.utilities.utils.LocalDateUtils.daysBetween
-import tmg.utilities.utils.LocalTimeUtils.fromTime
-import tmg.utilities.utils.LocalTimeUtils.isTimeValid
-import tmg.utilities.utils.LocalTimeUtils.requireFromTime
 
 internal class LocalDateUtilsTest {
 
-    @ParameterizedTest(name = "requireFromTime({0}) does now throw an exception")
+    @ParameterizedTest(name = "requireFromDate({0}) does now throw an exception")
     @CsvSource(
-        "14:00:00Z",
-        "14:00:00+0100",
-        "14:00:00",
-        "14:00:00.000",
-        "14:00"
+        "2020-1-1",
+        "2020-01-1",
+        "2020-01-01",
+        "2020-January-1",
+        "2020-Jan-01"
     )
-    fun `requireFromTime does not throw exception for valid time parses`(time: String) {
-        val expected = LocalTime.of(14, 0, 0)
-        assertEquals(expected, requireFromTime(time))
+    fun `requireFromDate does not throw exception for valid date parses`(date: String) {
+        val expected = LocalDate.of(2020, 1, 1)
+        assertEquals(expected, LocalDateUtils.requireFromDate(date))
     }
 
-    @ParameterizedTest(name = "requireFromTime({0}) is invalid and throws an exception")
+    @ParameterizedTest(name = "requireFromDate({0}) is invalid and throws an exception")
     @CsvSource(
-        "12",
-        "12:30:82",
-        "null",
+        "2020-1-32",
+        "1999-13-01",
+        "2023-12-32",
         "something-random"
     )
-    fun `requireFromTime throws exception for invalid date parses`(time: String) {
+    fun `requireFromDate throws exception for invalid date parses`(date: String) {
         assertThrows<DateTimeParseException> {
-            requireFromTime(time)
+            LocalDateUtils.requireFromDate(date)
         }
     }
 
-    @ParameterizedTest(name = "fromTime({0}) does now throw an exception")
+    @ParameterizedTest(name = "fromDate({0}) does now throw an exception")
     @CsvSource(
-        "14:00:00Z",
-        "14:00:00+0100",
-        "14:00:00",
-        "14:00:00.000",
-        "14:00"
+        "2020-1-1",
+        "2020-01-1",
+        "2020-01-01",
+        "2020-January-1",
+        "2020-Jan-01"
     )
-    fun `fromTime does not throw exception for valid date parses`(time: String) {
-        val expected = LocalTime.of(14, 0, 0)
-        assertEquals(expected, requireFromTime(time))
+    fun `fromDate does not throw exception for valid date parses`(date: String) {
+        val expected = LocalDate.of(2020, 1, 1)
+        assertEquals(expected, LocalDateUtils.fromDate(date))
     }
 
-    @ParameterizedTest(name = "fromTime({0}) is invalid and throws an exception")
+    @ParameterizedTest(name = "fromDate({0}) is invalid and throws an exception")
     @CsvSource(
-        "12",
-        "12:30:82",
-        "null",
+        "2020-1-32",
+        "1999-13-01",
+        "2023-12-32",
         "something-random"
     )
-    fun `fromTime returns null for invalid date`(time: String) {
-        assertNull(fromTime(time))
+    fun `fromDate returns null for invalid date`(date: String) {
+        assertNull(LocalDateUtils.fromDate(date))
     }
 
     @Test
-    fun `requireFromTime value maps correctly`() {
-        val input = "13:43:03"
-        val expected = LocalTime.of(13, 43, 3)
+    fun `requireFromDate value maps correctly`() {
+        val input = "1994-10-17"
+        val expected = LocalDate.of(1994, 10, 17)
 
-        assertEquals(expected, requireFromTime(input))
+        assertEquals(expected, LocalDateUtils.requireFromDate(input))
     }
 
     @Test
     fun `isDateValid returns true for valid date`() {
-        val input = "01:00:00"
-        assertTrue(isTimeValid(input))
+        val input = "2023-01-1"
+        assertTrue(LocalDateUtils.isDateValid(input))
     }
 
     @Test
     fun `isDateValid returns false for invalid date`() {
-        val input = "25:00:00"
-        assertFalse(isTimeValid(input))
-    }
-
-    @ParameterizedTest(name = "days between {0} and {1} is calculated to be {2}")
-    @CsvSource(
-        "2020-12-05,2020-12-04,-1",
-        "2020-12-05,2020-12-05,0",
-        "2020-12-05,2020-12-06,1",
-        "1995-12-05,2020-12-06,9133"
-    )
-    fun `daysBetween between two dates results in expected days`(start: String, end: String, days: Int) {
-        val inputStart = start.toLocalDate()!!
-        val inputEnd = end.toLocalDate()!!
-
-        assertEquals(days, daysBetween(inputStart, inputEnd))
+        val input = "2023-01-32"
+        assertFalse(LocalDateUtils.isDateValid(input))
     }
 }
